Fix problem again

+++ proyek-kelompok-8/FULLSTACK/peminjaman.php
@@ -2,12 +2,26 @@
 
 $peminjaman = pg_query($conn, "SELECT * FROM peminjaman");
 
-if (isset($_GET["nim"])) {
+if (isset($_GET["nim"]) && isset($_GET["kode"])) {
 	global $conn;
 
 	$nim = $_GET["nim"];
+	$kode = $_GET["kode"];
 
 	pg_query($conn, "DELETE FROM peminjaman WHERE nim = '$nim'");
+
+	$sepeda = pg_query($conn, "SELECT * FROM peminjaman WHERE kode_sepeda = '$kode'");
+
+	$jumlah = pg_num_rows($kode);
+
+	$jenisCheck = pg_query($conn, "SELECT * FROM sepeda WHERE kode = '$kode'");
+	$sepeda = pg_fetch_assoc($jenisCheck);
+
+	$tersedia = $sepeda["jumlah"] - $jumlah;
+
+	$updateSepeda = "UPDATE sepeda SET tersedia = $tersedia WHERE jenis = '$jenis'";
+
+	pg_query($conn, $updateSepeda);
 
 	header("Location: index.php?page=peminjaman");
 }
@@ -27,6 +41,7 @@ if (isset($_GET["nim"])) {
 		<thead>
 			<tr>
 				<th>NIM</th>
+				<th>Kode Sepeda</th>
 				<th>Jenis Sepeda</th>
 				<th>Tanggal Peminjaman</th>
 				<?php if ((isset($_SESSION["username"]))) : ?>
@@ -38,10 +53,11 @@ if (isset($_GET["nim"])) {
 			<?php while ($pmj = pg_fetch_assoc($peminjaman)) : ?>
 				<tr>
 					<td><?= $pmj['nim']; ?></td>
+					<td><?= $pmj['kode_sepeda']; ?></td>
 					<td><?= $pmj['jenis_sepeda']; ?></td>
 					<td><?= $pmj['tanggal_meminjam']; ?></td>
 					<?php if ((isset($_SESSION["username"]))) : ?>
-						<td><a href="index.php?page=peminjaman&nim=<?= $pmj['nim']; ?>">Kembailkan</a></td>
+						<td><a href="index.php?page=peminjaman&nim=<?= $pmj['nim']; ?>&kode=<?= $pmj['kode_sepeda']; ?>">Kembailkan</a></td>
 					<?php endif; ?>
 				</tr>
 			<?php endwhile; ?>
